refactor: 리뷰 내용 적용
 - 패키지 명 변경 time -> reservationtime
 - delete 메서드 명 변경 delete -> deleteById
 - usingRecursiveComparison 적용
