+++ src/main/java/roomescape/apply/reservation/application/ReservationRecorder.java
@@ -5,8 +5,8 @@ import roomescape.apply.reservation.domain.Reservation;
 import roomescape.apply.reservation.domain.repository.ReservationRepository;
 import roomescape.apply.reservation.ui.dto.ReservationRequest;
 import roomescape.apply.reservation.ui.dto.ReservationResponse;
-import roomescape.apply.time.application.ReservationTimeFinder;
-import roomescape.apply.time.domain.ReservationTime;
+import roomescape.apply.reservationtime.application.ReservationTimeFinder;
+import roomescape.apply.reservationtime.domain.ReservationTime;
 
 @Service
 public class ReservationRecorder {
