+++ src/main/java/roomescape/apply/reservationtime/application/ReservationTimeDeleter.java
@@ -1,7 +1,7 @@
-package roomescape.apply.time.application;
+package roomescape.apply.reservationtime.application;
 
 import org.springframework.stereotype.Service;
-import roomescape.apply.time.domain.repository.ReservationTimeRepository;
+import roomescape.apply.reservationtime.domain.repository.ReservationTimeRepository;
 
 import java.util.NoSuchElementException;
 
@@ -16,6 +16,6 @@ public class ReservationTimeDeleter {
 
     public void deleteReservationTimeBy(long id) {
         final long existId = reservationTimeRepository.checkIdExists(id).orElseThrow(NoSuchElementException::new);
-        reservationTimeRepository.delete(existId);
+        reservationTimeRepository.deleteById(existId);
     }
 }
