+++ src/main/java/roomescape/apply/reservationtime/application/ReservationTimeFinder.java
@@ -1,9 +1,9 @@
-package roomescape.apply.time.application;
+package roomescape.apply.reservationtime.application;
 
 import org.springframework.stereotype.Service;
-import roomescape.apply.time.domain.ReservationTime;
-import roomescape.apply.time.domain.repository.ReservationTimeRepository;
-import roomescape.apply.time.ui.dto.ReservationTimeResponse;
+import roomescape.apply.reservationtime.domain.ReservationTime;
+import roomescape.apply.reservationtime.domain.repository.ReservationTimeRepository;
+import roomescape.apply.reservationtime.ui.dto.ReservationTimeResponse;
 
 import java.util.List;
 import java.util.NoSuchElementException;
